Moved additional tests to server
